pyproject.toml
[tool.poetry]
name = "waherb"
version = "0.1.0"
description = "Staging database for WAHerb and WACensus data"
authors = ["Ashley Felton <ashley.felton@dbca.wa.gov.au>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.7"
Django = "3.0.7"
psycopg2-binary = "^2.8.5"
django-dotenv = "^1.4.2"
dj-database-url = "^0.5.0"
django-extensions = "^2.2.9"
dj-static = "^0.0.6"
gunicorn = "^20.0.4"
dbca-utils = "^1.1.1"
django-reversion = "^3.0.7"
django-mptt = "^0.11.0"
django-leaflet = "^0.26.0"
django_crispy_forms = "^1.9.1"
webtemplate-dbca = "^1.2.1"
django-bootstrap-pagination = "^1.7.1"

[tool.poetry.dev-dependencies]
ipython = "^7.15.0"
ipdb = "^0.13.2"
pygraphviz = "^1.5"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
